Refactor database manager to break cyclic imports and centralize utilities
- Extract download_content, calculate_hash, and TOML read/write helpers into a new
  surfactant.database_manager.utils module
- Remove in-file implementations and unused imports (hashlib, time, requests, tomlkit, etc.)
  from database_utils.py
- Update database_utils.py to import shared utilities from utils
- Change external_db_config.py to import download_content from utils at top-level
  (pylint: disable=import-outside-toplevel) instead of database_utils

This eliminates the R0401 cyclic-import between database_utils and external_db_config,
and consolidates shared functionality in a single utility module.
