+++ surfactant/database_manager/database_utils.py
@@ -18,7 +18,7 @@ from loguru import logger
 
 from surfactant.configmanager import ConfigManager
 from surfactant.database_manager.utils import (
-    download_content, save_db_version_metadata, calculate_hash, load_db_version_metadata
+    download_content, save_db_version_metadata, calculate_hash, load_db_version_metadata, get_source_for
     )
 
 
@@ -70,7 +70,6 @@ class BaseDatabaseManager(ABC):
         self.config = config
         # Attempt to retrieve an override URL using the database_dir (e.g., "js_library_patterns")
         # and the database_key (e.g., "retirejs").
-        from .external_db_config import get_source_for
 
         override_url = get_source_for(self.config.database_dir, self.config.database_key)
         if override_url:
