+++ surfactant/database_manager/utils.py
@@ -14,6 +14,7 @@ from typing import Any, Dict, Optional, Union
 from requests.exceptions import RequestException
 from pathlib import Path
 from loguru import logger
+import logging
 
 def download_content(url: str, timeout: int = 10, retries: int = 3) -> Optional[str]:
     """
@@ -163,3 +164,43 @@ def save_db_version_metadata(version_info: Union[str, Path], database_info: Dict
     }
     db_metadata.update(new_data)
     _write_toml_file(version_info, db_metadata)
+
+# URL for the hosted external TOML file on ReadTheDocs
+DEFAULT_EXTERNAL_DB_CONFIG_URL = (
+    "https://readthedocs.org/projects/surfacet-docs/downloads/latest/database_sources.toml"
+)
+
+
+def fetch_external_db_config(url: str = DEFAULT_EXTERNAL_DB_CONFIG_URL) -> dict:
+    """
+    Download and parse the external TOML file containing database source overrides.
+    Returns an empty dict on failure.
+    """
+    content = download_content(url)
+
+    try:
+        if content is not None:
+            config = tomlkit.parse(content)
+            return config
+    except tomlkit.exceptions.ParseError as e:
+        logging.warning("Error parsing TOML content: %s", e)
+    return {}
+
+
+def get_source_for(database_category: str, key: str) -> str:
+    """
+    Retrieve the URL for a given database category and key.
+
+    Args:
+        database_category (str): The category corresponding to a folder (e.g., 'js_library_patterns').
+        key (str): The specific key for the database (e.g., 'retirejs').
+
+    Returns:
+        str: The URL from the external configuration if available; otherwise, an empty string.
+    """
+    config = fetch_external_db_config()
+    try:
+        return config["sources"][database_category][key]
+    except KeyError:
+        logging.info("No external override found for [%s].%s", database_category, key)
+        return ""
